Add a border to all links, including <code><a>
The existing theme only had a link border on :hover, and no border even on
hover for links in a code span. This made it impossible to tell that a code
span was a link without hovering over it to see if the cursor changed.

I also changed the theme color from orange to blue.